updated derived solver and derived solver test

+++ calc.cpp
@@ -5,8 +5,8 @@
 int main(){
     using namespace std;
     DerivedSolver test;
-    Eigen::Matrix<int, 3, 3> a;
-    Eigen::Matrix<int, 3, 3> b;
+    Eigen::Matrix<float, 3, 3> a;
+    Eigen::Matrix<float, 3, 3> b;
     cout << test.solve(a, b) << endl;
     cout << test.verify(a);
     return 0;
